requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
sqlite3